+++ packages/core/src/components.d.ts
@@ -705,11 +705,15 @@ export namespace Components {
          */
         "dropShadow"?: boolean;
         /**
+          * For a green Highlight Box.
+         */
+        "green"?: boolean;
+        /**
           * To create a step indicator.
          */
         "step"?: number;
         /**
-          * For a while Highlight Box.
+          * For a white Highlight Box.
          */
         "white"?: boolean;
         /**
@@ -3406,11 +3410,15 @@ declare namespace LocalJSX {
          */
         "dropShadow"?: boolean;
         /**
+          * For a green Highlight Box.
+         */
+        "green"?: boolean;
+        /**
           * To create a step indicator.
          */
         "step"?: number;
         /**
-          * For a while Highlight Box.
+          * For a white Highlight Box.
          */
         "white"?: boolean;
         /**